Refactor logger
- Move to single logger
- Move logging middleware to avoid logging asset endpoints
