+++ moj-node/package.json
@@ -48,8 +48,6 @@
   "dependencies": {
     "axios": "0.19.2",
     "body-parser": "^1.19.0",
-    "bunyan": "^1.8.12",
-    "bunyan-format": "^0.2.1",
     "bunyan-request-logger": "^2.1.0",
     "cheerio": "^1.0.0-rc.3",
     "compression": "^1.7.4",
