+++ moj-node/server.js
@@ -1,7 +1,7 @@
 require('dotenv').config();
 
 const app = require('./server/index');
-const { logger } = require('./logger');
+const { logger } = require('./server/utils/logger');
 
 app.listen(app.get('port'), () => {
   logger.info(`Server listening on port ${app.get('port')}`);
