+++ moj-node/server/clients/hub.js
@@ -1,6 +1,6 @@
 const qs = require('querystring');
 const { baseClient } = require('./baseClient');
-const { logger } = require('../../logger');
+const { logger } = require('../utils/logger');
 const config = require('../config');
 
 const { getEstablishmentId } = require('../utils');
