+++ moj-node/server/middleware/auth.js
@@ -1,7 +1,7 @@
 const expressNTLM = require('express-ntlm');
 const { path } = require('ramda');
 const config = require('../config');
-const { logger } = require('../../logger');
+const { logger } = require('../utils/logger');
 
 const getOffenderNumberFrom = path(['user', 'offenderNo']);
