+++ moj-node/server/repositories/hubContent.js
@@ -2,7 +2,7 @@ const R = require('ramda');
 const qs = require('querystring');
 
 const config = require('../config');
-const { logger } = require('../../logger');
+const { logger } = require('../utils/logger');
 const { isEmpty, fillContentItems } = require('../utils');
 
 const {
